pyproject.toml
[project]
name = "benchling-mcp"
version = "0.1.1"
description = "MCP server for Benchling platform integration"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "eliot>=1.17.5",
    "pycomfort>=0.0.18",
    "fastmcp>=2.8.1",
    "fastapi>=0.115.13",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "benchling-sdk>=1.22.0",
    "python-dotenv>=1.0.0",
    "biopython>=1.83",
]

[project.scripts]
benchling-mcp = "benchling_mcp.server:cli_app_stdio"
server = "benchling_mcp.server:cli_app"
stdio = "benchling_mcp.server:cli_app_stdio"
sse = "benchling_mcp.server:cli_app_sse"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/benchling_mcp"]

[tool.hatch.build]
exclude = [
    "*.log",
    "logs/",
    "images/",
    ".venv/",
    ".pytest_cache/",
    "test/",
    "*.db",
    "temp*.sqlite*",
    "test*.sqlite*",
    "debug*.sqlite*",
]

[dependency-groups]
dev = [
    "just-agents-core>=0.7.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.12.0",
] 
